refactored caliber to hydra
